Rename CentralPlannerScheduler to Scheduler (#1781)
I find very little value in having that complicated name. In particular
it should be clear that the main class of scheduler.py is in fact
the Scheduler() class.

I've also renamed the tests testing the api to
test/scheduler_api_test.py. This should feel more intuitive to have
"scheduler" in the name of the file.

Former-commit-id: e02ba97c8d29e61cdd7d4c7a8845dcd05ed9a35b