README: Remove monthly downloads badge
The fact that there's only 4 downloads per month is not something we can really brag about. Let's just remove this bag. A historic note is that at some point that number fell from the tens of thousands downloads per month to just a handful. I suppose pypi fixed a bug on their side.

Former-commit-id: 83ff266f42f5b6ccd56e505cd177515bdcb96366