Adds BATCH_RUNNING to the visualiser (#1817)
In the earlier batching PR, I left out any visualiser elements. This
adds BATCH_RUNNING to the Task List selectors and Dependency Graph
legend. The legend needed to be widened a bit to accomodate the longer
name, and I replace commas with spaces while capitalizing the first
letter of each word in the legend.

Former-commit-id: f5bda8b36941b4937a53aba58454911e5412a88a