Allows clients to specify max_shown_tasks in task_list (#1847)
I have some scripts that process task_list for reporting. When the
response length goes beyond the maximum shown tasks, the scripts all
break. In order to allow the scripts to keep reporting as things get
backed up, this adds an optional argument to override the config's max
shown tasks in task_list, while keeping the configured default for the
visualizer.

Former-commit-id: ec485b7304ee990f0c4ed6af08960f11914be7f9