Fails batch_running tasks if worker dies early (#1850)
If a worker were to die between receiving batch tasks and responding with
the task that will run them, we want the tasks to fail. Our existing
methods to fail batch tasks don't work in this case, so I've extended the
method for failing RUNNING tasks to also work for BATCH_RUNNING in order
to handle this extra case.

Former-commit-id: 5a9cc05bc75c8ae1f2da36de74a95e38496b7686