Enables worker timeouts with only one worker
Worker timeouts rely on running the task in a separate process. In order
to get them to work with a single worker, we need to run tasks that have
timeouts in separate processes. This change does that while still allowing
tasks that don't have a timeout set to run in the main process.


Former-commit-id: afe01d32eb8a7fac1249554a8a658393a66a530e