Tries using path for non-hdfs targets in hadoop_jar (#1824)
When writing a map-reduce job that takes S3 inputs, I found that the
arguments were incorrect. As S3Target does not inherit from HdfsTarget,
the code goes to the base case of calling str. To be correct a bit more
often in this case, I've adjusted the code to try checking the path
attribute first.

Former-commit-id: aa30be84ff1a5d809fd12dbf49d6c2afdfbce95b