Refine disabling of workers (#1839)
This commit's PR started small but grew quite big. But there's two major changes:

 * Bugfix: don't give tasks to disabled assistants

Previously assistants would keep getting tasks and never stop. Despite
that a user have clicked "disable worker" in the WebUI.

 * Invent a worker-state, pass it along in get_work and worker_list

This gives an opportunity for workers to know if the scheduler considers them disabled. It also allows the WebUI to show if a worker is disabled or not. Making the UI experience ever so more responsive. 

The old "greyed out worker" now means will mean "disabled". Previously it was equivalent to "has no pending tasks". But because we also associate "greyed out" with "can't click disable", the WebUI have
confused at least me for a while.

Former-commit-id: cac500a1f480de32e603b6421723c87b4f4fc5ad