Adds a flag to turn off failure e-mails in the worker
This is intended to be used in conjunction with batch notification e-mails
in the scheduler, allowing the types of e-mails that will be sent to the
scheduler to not be sent by workers as well. This can be preferable to
simply disabling e-mail because it still allows some types of e-mails
to be sent and also works when the scheduler and worker are on the same
machine.


Former-commit-id: 789368e29b50fc801210abf3b2e35df363416488