Prevent changing resources from exceeding limits
The scheduler enforces resource constraints during get_work by adding up
the resources of currently running jobs and making sure that no new jobs
push any resource beyond its limit. This normally works fine, but can
run into problems when jobs change their resources after being returned
by get_work. This can happen in two ways.

The first way this can happen is that another worker schedules the job
while it's running with different resources. We want this change to be
made, as the new resources are what should be associated with the job
should it fail and need to re-run.

The second way this can happen is if a worker gets the job as part of
a batch run and returns a task with different resources.

The solution to both of these cases is to simply store the resources
that were allocated to the task during get_work and use those to count
the currently running resources. In both cases, this prevents us from
running an arbitrary number of resource-constrained tasks by switching
the resources after the limits are done being applied. We can still get
more tasks running than intended due to the mixed state but we at least
preserve some finite limit.


Former-commit-id: d8310bdfc051b60ffa8e677be90951c35b06c6e1