LocalTarget: Allow moves across devices (#1877)
With `os.rename` you get a `OSError: [Errno 18] Invalid cross-device link`
if you attempt to move a file across filesystems.

If the destination is on a different filesystem,
we approximate atomicity by first copying the source
to a temporary file on the target filesystem, followed
by a rename. Finally the source file is removed.

Former-commit-id: e9ce79e02897d993012c448323e797f084f5b7f8