Make temporary_path() decorator create directories (#1919)
I noticed that my program crashed in the cases where I didn't manually
create the directories. And in all cases so far I've needed for the
directories to be created.

Will I definetly think this feature should be possible to turn off, I
think it's best to have as "on by default". But I left it out for anyone
making a patch in the future.

Former-commit-id: 52f0e49c7f17080450cfbec4483be1aa4fda8f7d