namespaces: Fix changed behavior since #1953
This patch kinda rethinks and modernizes all the metamagic
between the Task class and the Register meta class. I think the less the
better. The logic inside the Register has shrunk and been replaced with
more easy to read logic in the Task class.

Furthermore. It identifies a code reuse opportunity for the luigi.util
module and also fixes a bug in the util-decorators that we love and use.

Overall, after this patch, luigi has become even more functional and
contains less if-and-but logic for how to determine the namespaces. Also
eventually there will be no strange magic of why the task_family
isn't assignable but task_namespace is.


Former-commit-id: d40c7535ddee75738febca8b5b1bf4f76975231d