externalize: Dont register the copied tasks (#1975)
This fixes an implementation detail, so that luigi.task.externalize
actually does what it says it does in the documentation.  Hence I
hope pretty uncontroversial. This makes it so that calling externalize
don't register a new task. A pretty serious bug.

Now, there's still a caveat, as we're mostly hiding that we do side
affects, if anyone would actually inherit from what is returned by
externalize.  That subtask wouldn't be registered either.

This code is quite easy to get wrong somewhere. Luckily luigi is
decently tested. To make me even more confident I also added a couple of
tests:

 * Two tests in task_test.py that didn't pass before.
 * I created the task_register_test.py file.
 * I added some util tests too, because they are also all about
   Copying classes and namespace clashes. And I want to be on the safe
   side.
Former-commit-id: f7e0b7710fd8b12f27625b1efb62a8fde6e206c4