Make temporary_path work with paths not containing a directory (#2058)
The context manager returned by FileSystemTarget.temporary_path creates
the parent directory of the target path. However this fails when the
path does not contain a directory and an exception is raised. To be able
to use paths without a directory (e.g. "foobar.txt"), the creation of
the parent directoy is skipped in this case.

A test case which tests the above scenario has been added.

Former-commit-id: ffaedb337b0d0355a9255bf28c0ae57fb71c856a