Remove search query from url bar on empty query
When emptying the search query box, we don't trigger the change to the
query in the URL bar, resulting in a value being stored in the URL
without a corresponding search in the box. To fix this, we simply delete
the search__search key when the search box is empty.


Former-commit-id: 20ab288a5b8cef3e62081eaf8c1dbfa2e596eed2