Update individual resource counts from the vis
Sometimes you need to quickly tweak a resource in response to your jobs
falling behind or your cluster getting overloaded. In order to make
such tweaks a little easier, this adds the option to change the limits
for all resources stored in the scheduler right in the visualiser.

This works exactly like updating the worker counts in the workers tab,
except that the counts are referenced in the progress bar rather than on
the button itself. This is because we already have the counts in the
progress bar, so adding them to a button would only complicate things.

The only other difference between this and the worker counts is that
this allows resources to be set to 0, effectively pausing certain kinds
of jobs.

Note that when you update the resource counts using the visualiser, it
only affects the currently running scheduler. If you restart the
scheduler it will reload the resource counts from the config.


Former-commit-id: 1b8de2944d4ca2efe1937030ef9bc044ed519e66