Move visualiser graph type default to a constant
In my fork of Luigi, I use svg as the graph type default. This results
in annoying merges when javascript related to handling the defaults
changes. In order to make different defaults easier to maintain, this
uses a constant in visualiserApp.js to determine which graph type is
default rather than hard-coding it everywhere the default value is used.

This also cleans up a few places where the default was being set
unnecessarily:
  - don't set it in html since we set it in javascript instantly anyway
  - remove code to set buttons to checked since this has no meaning

Former-commit-id: 3a8f381e4fee0cff431d8d1e7c8e94022251ec71