Preserve graph options when clicking d3 nodes (#2139)
When clicking a d3 node, the new url was built up with a static set of
options, clearing out things like hide done and invert. This means the
graph would flip or gain lots of nodes when you click a link. In order
to better preserve the options, we now simply change the taskId in the
current URL.

Former-commit-id: d5c193b0f33c945531e9f07525d0cab72ee9494d